compile for macos

+++ editor/highlighter.h
@@ -33,7 +33,7 @@ struct span_info_t {
 
 struct bracket_info_t {
     size_t line;
-    int char_idx;
+    size_t char_idx;
     int bracket;
     bool open;
     bool unpaired;
